backend/requirements.txt
Django==5.1.4
Pillow==11.0.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
django-cors-headers==4.6.0 
openpyxl==3.1.2 
mysqlclient==2.2.4
reportlab==4.1.0 